pyproject.toml
[tool.poetry]
name = "stx-blog"
version = "0.1.0"
description = ""
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.2.0"
pytest-playwright = "^0.3.0"
pytest-base-url = "^2.0.0"
pre-commit = "^3.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
